Include chrome SConscript.unit_tests in Linux build.
BUG=2335

Original patch by phajdan.jr@gmail.com in
http://codereview.chromium.org/2891

R=mark,sgk
Review URL: http://codereview.chromium.org/2939

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2319 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/safe_browsing/chunk_range_unittest.cc
@@ -49,7 +49,7 @@ TEST(SafeBrowsingChunkRangeTest, TestChunksToRanges) {
   chunks.push_back(4);
   chunks.push_back(7);
   ChunksToRanges(chunks, &ranges);
-  EXPECT_EQ(ranges.size(), 2);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(2));
   EXPECT_EQ(ranges[0].start(), 1);
   EXPECT_EQ(ranges[0].stop(),  4);
   EXPECT_EQ(ranges[1].start(), 7);
@@ -68,7 +68,7 @@ TEST(SafeBrowsingChunkRangeTest, TestChunksToRanges) {
   chunks.push_back(9);
   chunks.push_back(10);
   ChunksToRanges(chunks, &ranges);
-  EXPECT_EQ(ranges.size(), 1);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(1));
   EXPECT_EQ(ranges[0].start(), 3);
   EXPECT_EQ(ranges[0].stop(),  10);
 
@@ -85,7 +85,7 @@ TEST(SafeBrowsingChunkRangeTest, TestChunksToRanges) {
   chunks.push_back(15);
   chunks.push_back(17);
   ChunksToRanges(chunks, &ranges);
-  EXPECT_EQ(ranges.size(), 8);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(8));
 
   chunks.clear();
   ranges.clear();
@@ -93,7 +93,7 @@ TEST(SafeBrowsingChunkRangeTest, TestChunksToRanges) {
   // Test a single chunk number.
   chunks.push_back(17);
   ChunksToRanges(chunks, &ranges);
-  EXPECT_EQ(ranges.size(), 1);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(1));
   EXPECT_EQ(ranges[0].start(), 17);
   EXPECT_EQ(ranges[0].stop(),  17);
 
@@ -111,7 +111,7 @@ TEST(SafeBrowsingChunkRangeTest, TestChunksToRanges) {
   chunks.push_back(7);
   chunks.push_back(7);
   ChunksToRanges(chunks, &ranges);
-  EXPECT_EQ(ranges.size(), 2);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(2));
   EXPECT_EQ(ranges[0].start(), 1);
   EXPECT_EQ(ranges[0].stop(), 3);
   EXPECT_EQ(ranges[1].start(), 7);
@@ -124,7 +124,7 @@ TEST(SafeBrowsingChunkRangeTest, TestStringToRanges) {
 
   std::string input = "1-100,398,415,1138-2001,2019";
   EXPECT_TRUE(StringToRanges(input, &ranges));
-  EXPECT_EQ(ranges.size(), 5);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(5));
   EXPECT_EQ(ranges[0].start(), 1);
   EXPECT_EQ(ranges[0].stop(),  100);
   EXPECT_EQ(ranges[1].start(), 398);
@@ -136,13 +136,13 @@ TEST(SafeBrowsingChunkRangeTest, TestStringToRanges) {
 
   input = "1,2,3,4,5,6,7";
   EXPECT_TRUE(StringToRanges(input, &ranges));
-  EXPECT_EQ(ranges.size(), 7);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(7));
 
   ranges.clear();
 
   input = "300-3001";
   EXPECT_TRUE(StringToRanges(input, &ranges));
-  EXPECT_EQ(ranges.size(), 1);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(1));
   EXPECT_EQ(ranges[0].start(),  300);
   EXPECT_EQ(ranges[0].stop(),  3001);
 
@@ -150,7 +150,7 @@ TEST(SafeBrowsingChunkRangeTest, TestStringToRanges) {
 
   input = "17";
   EXPECT_TRUE(StringToRanges(input, &ranges));
-  EXPECT_EQ(ranges.size(), 1);
+  EXPECT_EQ(ranges.size(), static_cast<size_t>(1));
   EXPECT_EQ(ranges[0].start(), 17);
   EXPECT_EQ(ranges[0].stop(),  17);
 
@@ -169,7 +169,7 @@ TEST(SafeBrowsingChunkRangeTest, TestRangesToChunks) {
   std::vector<int> chunks;
   RangesToChunks(ranges, &chunks);
 
-  EXPECT_EQ(chunks.size(), 5);
+  EXPECT_EQ(chunks.size(), static_cast<size_t>(5));
   EXPECT_EQ(chunks[0], 1);
   EXPECT_EQ(chunks[1], 2);
   EXPECT_EQ(chunks[2], 3);
